+++ core/model/field/bet_area/required/field.go
@@ -1,17 +1,17 @@
 package required
 
 type IDField struct {
-	ID int64 `json:"id" gorm:"column:id"` // 注區 id
+	ID int `json:"id" gorm:"column:id"` // 注區 id
 }
 type GameIdField struct {
-	GameId int64 `json:"game_id" gorm:"column:game_id"` // 遊戲id(關聯)
+	GameId int `json:"game_id" gorm:"column:game_id"` // 遊戲id(關聯)
 }
 type NameField struct {
 	Name string `json:"name" gorm:"column:name"` // 注區名稱
 }
 type MinLimitField struct {
-	MinLimit int64 `json:"min_limit" gorm:"column:min_limit"` // 最小限額
+	MinLimit int `json:"min_limit" gorm:"column:min_limit"` // 最小限額
 }
 type MaxLimitField struct {
-	MaxLimit int64 `json:"max_limit" gorm:"column:max_limit"` // 最大限額
+	MaxLimit int `json:"max_limit" gorm:"column:max_limit"` // 最大限額
 }
