+++ core/model/field/game/required/field.go
@@ -1,7 +1,7 @@
 package required
 
 type IDField struct {
-	ID int64 `json:"id" gorm:"column:id"` // 遊戲id
+	ID int `json:"id" gorm:"column:id"` // 遊戲id
 }
 type NameField struct {
 	Name string `json:"name" gorm:"column:name"` // 遊戲名稱
