+++ core/model/field/game/required/field_pk.go
@@ -1,5 +1,5 @@
 package required
 
 type GameIDField struct {
-	GameID int64 `json:"game_id" gorm:"column:game_id"` // 遊戲id
+	GameID int `json:"game_id" gorm:"column:game_id"` // 遊戲id
 }
