+++ core/model/field/game_status/optional/field.go
@@ -1,15 +1,15 @@
 package optional
 
 type GameIDField struct {
-	GameID *int64 `json:"game_id,omitempty" gorm:"column:game_id"` // 遊戲id
+	GameID *int `json:"game_id,omitempty" gorm:"column:game_id"` // 遊戲id
 }
 
 type CountDownField struct {
-	CountDown *int32 `json:"count_down,omitempty" gorm:"column:count_down"` // 當前階段時長
+	CountDown *int `json:"count_down,omitempty" gorm:"column:count_down"` // 當前階段時長
 }
 
 type StageField struct {
-	Stage *int32 `json:"stage,omitempty" gorm:"column:stage"` // 階段(投注=1/發牌=3/結算=2)
+	Stage *int `json:"stage,omitempty" gorm:"column:stage"` // 階段(投注=1/發牌=3/結算=2)
 }
 
 type RoundInfoIDField struct {
@@ -17,5 +17,5 @@ type RoundInfoIDField struct {
 }
 
 type DeckRoundField struct {
-	DeckRound *int32 `json:"deck_round,omitempty" gorm:"column:deck_round"` // 局數
+	DeckRound *int `json:"deck_round,omitempty" gorm:"column:deck_round"` // 局數
 }
