+++ core/model/field/odd/required/field.go
@@ -1,13 +1,13 @@
 package required
 
 type IDField struct {
-	ID int64 `json:"id" gorm:"column:id"` // id
+	ID int `json:"id" gorm:"column:id"` // id
 }
 type GameIdField struct {
-	GameId int64 `json:"game_id" gorm:"column:game_id"` // 遊戲id(關聯)
+	GameId int `json:"game_id" gorm:"column:game_id"` // 遊戲id(關聯)
 }
 type BetAreaIdField struct {
-	BetAreaId int64 `json:"bet_area_id" gorm:"column:bet_area_id"` // 注區id(關聯)
+	BetAreaId int `json:"bet_area_id" gorm:"column:bet_area_id"` // 注區id(關聯)
 }
 type OddField struct {
 	Odd float32 `json:"odd" gorm:"column:odd"` // 賠率
